Add embed migrations

+++ internal/transport/http/middleware/auth_middleware.go
@@ -55,7 +55,7 @@ func (m *AuthMiddleware) HandleFunc(next http.HandlerFunc) http.HandlerFunc {
 			return
 		}
 
-		user, err := m.UserService.GetUserById(ctx, authorizedUserID)
+		user, err := m.UserService.GetUserById(ctx, uint64(authorizedUserID))
 		if err != nil || user == nil || user.ID != token.UserID {
 			http.Error(w, http.StatusText(http.StatusUnauthorized), http.StatusUnauthorized)
 		}
